Apply public release runtime fixes

+++ server/session.py
@@ -32,8 +32,8 @@ class Session:
     eval_sql: str = ""
     condition_at: Optional[float] = None
     speed_factor: float = 1.0
-    contact_get_time: Optional[int] = None
-    contact_post_time: Optional[int] = None
+    contact_get_time: Optional[float] = None
+    contact_post_time: Optional[float] = None
     contact_message: Optional[str] = None
     start_page: Optional[str] = None
